% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adams.R
\docType{data}
\name{Adams}
\alias{Adams}
\title{The meta-analysis by Adams et al. (2018) to synthesize the effects of fluvastain 20mg/day
on high-density lipoprotein.}
\format{A data frame with 32 rows and 2 variables:
\describe{
  \item{y}{mean differences}
  \item{s2}{the corresponding within-study variance}
}}
\source{
\url{https://www.cochranelibrary.com/cdsr/doi/10.1002/14651858.CD012282.pub2/epdf/full}
}
\usage{
Adams
}
\description{
A dataset containing 32 studies, with the number of patients in treated and control groups.
The number of events for each group is also reported for each study. The log odds ratio and
its corresponding variance are reported or caculated for each study.
}
\keyword{datasets}
