% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_lambda.R
\name{determine.lambda}
\alias{determine.lambda}
\title{Obtain candidate values for the tuning parameter lambda}
\usage{
determine.lambda(est.eff, est.var, penalty = "tau2", n.lambda = 100,
  lambda.scale = "log", tol, lam.c = 1.2)
}
\arguments{
\item{est.eff}{an observed effect size vector}

\item{est.var}{the corresponding within-study varaince vector}

\item{penalty}{the penalty function used in the penalized likelihood, the default value is 'tau2'}

\item{n.lambda}{the number of candidate values of lambda, the default number is 100}

\item{lambda.scale}{the transformed scale of lambda, the default value is 'log'}

\item{tol}{the relative convergence tolerence in optimization}

\item{lam.c}{a scalar used to mutiply lambda_max, by controlling its magnitude we can specify the upper bound of candidate values for lambda}
}
\value{
a threshold value for tuning parameters lambda
}
\description{
Obtain candidate values of the tuning parameter lambda
}
\details{
A threshold value of lambda is used to define the upper bound of candidate values for lambda, and the lower bound is zero
}
\author{
Yipeng
}
