% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_original_effects.R
\name{forest_original_effects}
\alias{forest_original_effects}
\title{Plot a forestplot including the penalization methods}
\usage{
forest_original_effects(y, s2, out, adjust_font_size = 0, llimit = -10,
  ulimit = 20, user_define_x_ticks = FALSE, x_ticks,
  effect_type = "MD")
}
\arguments{
\item{y}{an observed effect size vector}

\item{s2}{the corresponding within-study varaince vector}

\item{out}{the summary outcome of the metagen( ) in the R package meta}

\item{adjust_font_size}{a scalar used to control the font size of labels and ticks}

\item{llimit}{the lower limit for clipping confidence intervals to arrows}

\item{ulimit}{the upper limit for clipping confidence intervals to arrows}

\item{user_define_x_ticks}{if it is true, then user-specified x-axis tick marks are used. Otherwise, default ticks are used}

\item{x_ticks}{the user-specified x-axis tick marks, it must be provided when user_define_x_ticks is TRUE. x_ticks is not specified when user_define_x_ticks is FALSE, and the plot will use the default values}

\item{effect_type}{the type of effect sizes, which can be 'MD' (mean difference), 'SMD' (standardized mean difference), 
or 'RD' (risk difference)}
}
\description{
Produces a forest plot of a given meta-analysis when effects are continous outcomes
}
\details{
The result plot summarises the point estimate and the 95% confidecne interval for the overall effect size of four models in meta-analysis: the common-effect model, the random-effects model, the penalization method by tuning lambda, and the penalization method by tuning tau. The effect sizes of continuous outcomes are consided in this function: mean differences, standardized mean differences, and risk differences
}
\author{
Yipeng
}
