% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_reverse_log.R
\name{forest_reverse_log}
\alias{forest_reverse_log}
\title{Plot a forestplot including the penalization methods}
\usage{
forest_reverse_log(y, s2, out, adjust_font_size = 0, llimit = 0.25,
  ulimit = 16, user_define_x_ticks = FALSE, x_ticks,
  effect_type = "OR")
}
\arguments{
\item{y}{an observed effect size vector that is measured on log scale}

\item{s2}{the corresponding within-study varaince vector}

\item{out}{the summary outcome of the metagen( ) in the R package meta}

\item{adjust_font_size}{a scalar used to control the font size of labels and ticks}

\item{llimit}{the lower limit for clipping confidence intervals to arrows}

\item{ulimit}{the upper limit for clipping confidence intervals to arrows}

\item{user_define_x_ticks}{if it is true, then user-specified x-axis tick marks are used. Otherwise, default ticks are used}

\item{x_ticks}{the user-specified x-axis tick marks, it must be provided when user_define_x_ticks is TRUE. x_ticks is not specified when user_define_x_ticks is FALSE, and the plot will use the default values}

\item{effect_type}{the type of effect sizes that are not on the log scale, which can be 'OR' (odds ratio)
or 'RR' (relative risk)}
}
\description{
Produce a forest plot of a given meta-analysis when effects are binary outcomes
}
\details{
The result plot summarises the point estimate and the 95% confidecne interval for the overall effect size of four models in meta-analysis: the common-effect model, the random-effects model, the penalization method by tuning lambda, and the penalization method by tuning tau. The effect sizes of binary outcomes are consided in this function: log odds ratio and log relative risk. This function reverses the results of effects on log scale to orignal scale
}
\author{
Yipeng
}
