% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaml.R
\name{metaml}
\alias{metaml}
\title{Compute the ML estimate for the between-study variance}
\usage{
metaml(y, s2, tol = 10^(-10))
}
\arguments{
\item{y}{an observed effect size vector}

\item{s2}{the corresponding within-study varaince vector}

\item{tol}{the relative convergence tolerence in optimization}
}
\value{
the ML estimate of the between-study variance
}
\description{
Obtain the ML estimate of the between-study variance
}
\details{
For a given meta-analysis, optimize the log-likelihood function to get the ML estimate of the between-study variance
}
\author{
Yipeng
}
