% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaoutliers_plot.R
\name{metaoutliers.plot}
\alias{metaoutliers.plot}
\title{A study-specific standardized residuals plot of a meta-analysis}
\usage{
metaoutliers.plot(est.eff, est.var)
}
\arguments{
\item{est.eff}{an observed effect size vector}

\item{est.var}{the corresponding within-study varaince vector}
}
\description{
Plot study-specific standardized residuals for a given meta-analysis to check potential outliers
}
\author{
Yipeng
}
