% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapen_lamb.R
\name{metapen.lamb}
\alias{metapen.lamb}
\title{Obtain all necessary results of the penalization method by tuning lambda}
\usage{
metapen.lamb(est.eff, est.var, tau2.re, penalty = "tau2", lambda,
  n.lambda, lambda.scale = "log", tau2.ml, tol = 10^(-10),
  lam.c = 1.2)
}
\arguments{
\item{est.eff}{an observed effect size vector}

\item{est.var}{the corresponding within-study varaince vector}

\item{tau2.re}{the estimated between-study variance of the random-effects model, the default value is the ML estimate}

\item{penalty}{the penalty function used in the penalized likelihood, the default value is 'tau2'}

\item{lambda}{a vector of candidate values for lambda}

\item{n.lambda}{the number of candidate values of lambda, the default number is 100}

\item{lambda.scale}{the scale of lambda will display, 'log' represent the log scale, 'linear' represent the linear scale}

\item{tau2.ml}{the ML estimate of the between-study variance. If it is not specified, the function will automatically compute the value}

\item{tol}{the relative convergence tolerence in optimization}

\item{lam.c}{a scalar used to adjust the upper bound of candidate values for lambda}
}
\value{
a list consist of the results of the fixed-effect model, the random-effects model and the penalization approach by tuning lambda
}
\description{
Conduct the penalization method by tuning lambda
}
\details{
Many values are returned in the outcome list. n.study is the number of studies for a given meta-analysis; tau2.re is the estimated between-study variance of the random-effects model; I2 is the value of the I^2 statistic; mu.fe represents the overall effect size estimate of the common-effect model; se.fe is the estimate of the standard deviation of the common-effect model; mu.re is the overall effect size estimate of the random-effects model; se.re is the estimated standard deviation of the random-effects model; tau2 (tau) is a vector of the between-study variance (standard deviation) estimates correspond to candidate values of lambda; loss is a vector of losses for the loss function of lambda; tau.opt (tau2.opt) is the optimal between-study standard deviation (variance) using the penalization method by tuning lambda; mu.opt and se.opt are the overall effect size and the between-study standard deviation correspond to optimal lambda
}
\author{
Yipeng
}
