% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapen_lambda.R
\name{metapen.lambda}
\alias{metapen.lambda}
\title{Implement the penalization method by tuning lambda}
\usage{
metapen.lambda(est.eff, est.var, penalty = "tau2", lambda, tau2.ml,
  tol = 10^(-10))
}
\arguments{
\item{est.eff}{an observed effect size vector}

\item{est.var}{the corresponding within-study variance vector}

\item{penalty}{the penalty function used in the penalized likelihood, the default value is 'tau2'}

\item{lambda}{a vector of candidate values for lambda}

\item{tau2.ml}{the ML estimate of the between-study variance}

\item{tol}{the relative convergence tolerence in optimization}
}
\value{
a list consist of candidate values of lambda, estimates of the corresponding overall effect size and the between-study variance
}
\description{
Compute the corresponding overall effect size estimates and the between-study variances for a vector of lambda
}
\details{
For a vector of candidate values of lambda, calculate estimates for the overall effect size and the between-study varaince
}
\author{
Yipeng
}
