% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapen_lambdas_cv.R
\name{metapen.lambdas.cv}
\alias{metapen.lambdas.cv}
\title{Obtain values to show relationships among different parameters in the penalization method by tuning lambda}
\usage{
metapen.lambdas.cv(est.eff, est.var, penalty = "tau2", lambda, n.lambda,
  lambda.scale = "log", tau2.ml, tol = 10^(-10), lam.c = 1.2)
}
\arguments{
\item{est.eff}{an observed effect size vector}

\item{est.var}{the corresponding within-study variance vector}

\item{penalty}{the penalty function used in the penalized likelihood, the default value is 'tau2'}

\item{lambda}{a vector of candidate values for lambda}

\item{n.lambda}{the number of elements in the lambda vector}

\item{lambda.scale}{the scale of lambda will display, 'log' represent the log scale, 'linear' represent the linear scale}

\item{tau2.ml}{the ML estimate of the between-study variance. If it is not specified, the function will automatically compute the value}

\item{tol}{the relative convergence tolerence in optimization}

\item{lam.c}{a scalar used to adjust the upper bound of candidate values for lambda}
}
\value{
values of different parameters of the penalization approach by tuning lambda
}
\description{
This function is used to save values to visualize the cross-validation process of the penalization method by tuning lambda
}
\details{
All augements are the same in metapen.lamb function
}
\author{
Yipeng
}
