% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapen_tau.R
\name{metapen.tau}
\alias{metapen.tau}
\title{Obtain the results of the penalization method by tuning tau}
\usage{
metapen.tau(y, s2, tau2.re, upp, n.cand = 100, tol = 10^(-10))
}
\arguments{
\item{y}{an observed effect size vector}

\item{s2}{the corresponding within-study varaince vector}

\item{tau2.re}{the between-study variance estimate of the random-effects model, if it is not specified,
the ML estimate will be used as default}

\item{upp}{a scalar used to multiply tau2.re. Controlling its magnitude can adjust the range of tau}

\item{n.cand}{the number of candidates for the standard deviation considered in the penalization method}

\item{tol}{the relative convergence tolerence in optimization}
}
\value{
values of different parameters of the penalization approach by tuning tau
}
\description{
This function is used to obtain the results of the penalization method by tuning tau
}
\details{
Many values are returned as outcomes, which are similar as the results of the function metapen_lamb( ). Except two values, tau.cand represent the candidates of standard deviations used in the penalization method, where tau2.cand represent the corresponding candidates of between-study varainces
}
\author{
Yipeng
}
